+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - CrestronFusion - Added RebuildRVI Console Command
+ - CrestronFusion - Cleaned up RVI generation logging, set finished generating log as a notice
 
 ## [5.5.1] - 2020-10-12
 ### Changed